+++ hotdata/models/refresh_response.py
@@ -20,13 +20,12 @@ from pydantic import BaseModel, ConfigDict, Field, StrictStr, ValidationError, f
 from typing import Any, List, Optional
 from hotdata.models.connection_refresh_result import ConnectionRefreshResult
 from hotdata.models.schema_refresh_result import SchemaRefreshResult
-from hotdata.models.submit_job_response import SubmitJobResponse
 from hotdata.models.table_refresh_result import TableRefreshResult
 from pydantic import StrictStr, Field
 from typing import Union, List, Set, Optional, Dict
 from typing_extensions import Literal, Self
 
-REFRESHRESPONSE_ONE_OF_SCHEMAS = ["ConnectionRefreshResult", "SchemaRefreshResult", "SubmitJobResponse", "TableRefreshResult"]
+REFRESHRESPONSE_ONE_OF_SCHEMAS = ["ConnectionRefreshResult", "SchemaRefreshResult", "TableRefreshResult"]
 
 class RefreshResponse(BaseModel):
     """
@@ -38,10 +37,8 @@ class RefreshResponse(BaseModel):
     oneof_schema_2_validator: Optional[TableRefreshResult] = None
     # data type: ConnectionRefreshResult
     oneof_schema_3_validator: Optional[ConnectionRefreshResult] = None
-    # data type: SubmitJobResponse
-    oneof_schema_4_validator: Optional[SubmitJobResponse] = None
-    actual_instance: Optional[Union[ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult]] = None
-    one_of_schemas: Set[str] = { "ConnectionRefreshResult", "SchemaRefreshResult", "SubmitJobResponse", "TableRefreshResult" }
+    actual_instance: Optional[Union[ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult]] = None
+    one_of_schemas: Set[str] = { "ConnectionRefreshResult", "SchemaRefreshResult", "TableRefreshResult" }
 
     model_config = ConfigDict(
         validate_assignment=True,
@@ -79,17 +76,12 @@ class RefreshResponse(BaseModel):
             error_messages.append(f"Error! Input type `{type(v)}` is not `ConnectionRefreshResult`")
         else:
             match += 1
-        # validate data type: SubmitJobResponse
-        if not isinstance(v, SubmitJobResponse):
-            error_messages.append(f"Error! Input type `{type(v)}` is not `SubmitJobResponse`")
-        else:
-            match += 1
         if match > 1:
             # more than 1 match
-            raise ValueError("Multiple matches found when setting `actual_instance` in RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult. Details: " + ", ".join(error_messages))
+            raise ValueError("Multiple matches found when setting `actual_instance` in RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult. Details: " + ", ".join(error_messages))
         elif match == 0:
             # no match
-            raise ValueError("No match found when setting `actual_instance` in RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult. Details: " + ", ".join(error_messages))
+            raise ValueError("No match found when setting `actual_instance` in RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult. Details: " + ", ".join(error_messages))
         else:
             return v
 
@@ -122,19 +114,13 @@ class RefreshResponse(BaseModel):
             match += 1
         except (ValidationError, ValueError) as e:
             error_messages.append(str(e))
-        # deserialize data into SubmitJobResponse
-        try:
-            instance.actual_instance = SubmitJobResponse.from_json(json_str)
-            match += 1
-        except (ValidationError, ValueError) as e:
-            error_messages.append(str(e))
 
         if match > 1:
             # more than 1 match
-            raise ValueError("Multiple matches found when deserializing the JSON string into RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult. Details: " + ", ".join(error_messages))
+            raise ValueError("Multiple matches found when deserializing the JSON string into RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult. Details: " + ", ".join(error_messages))
         elif match == 0:
             # no match
-            raise ValueError("No match found when deserializing the JSON string into RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult. Details: " + ", ".join(error_messages))
+            raise ValueError("No match found when deserializing the JSON string into RefreshResponse with oneOf schemas: ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult. Details: " + ", ".join(error_messages))
         else:
             return instance
 
@@ -148,7 +134,7 @@ class RefreshResponse(BaseModel):
         else:
             return json.dumps(self.actual_instance)
 
-    def to_dict(self) -> Optional[Union[Dict[str, Any], ConnectionRefreshResult, SchemaRefreshResult, SubmitJobResponse, TableRefreshResult]]:
+    def to_dict(self) -> Optional[Union[Dict[str, Any], ConnectionRefreshResult, SchemaRefreshResult, TableRefreshResult]]:
         """Returns the dict representation of the actual instance"""
         if self.actual_instance is None:
             return None